Handle edge case...
The program was not correctly counting the game as complete if the puzzle had more than one word as the space was not being accounted for. Refactored calculateStatus() method to automatically count a space so the user did not have to guess the space character in order to satisfy the win condition.